+++ kurtosis-module/impl/participant_network/cl/lodestar/lodestar_launcher.go
@@ -3,6 +3,7 @@ package lodestar
 import (
 	"fmt"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/availability_waiter"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/cl_client_rest_client"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator"
@@ -10,7 +11,6 @@ import (
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
 	"github.com/kurtosis-tech/stacktrace"
-	"github.com/sirupsen/logrus"
 	"path"
 	"time"
 )
@@ -29,15 +29,11 @@ const (
 	discoveryPortNum uint16 = 9000
 	httpPortNum             = 4000
 
-
 	genesisConfigYmlRelFilepathInSharedDir = "genesis-config.yml"
 	genesisSszRelFilepathInSharedDir       = "genesis.ssz"
 
 	maxNumHealthcheckRetries      = 20
 	timeBetweenHealthcheckRetries = 1 * time.Second
-
-	maxNumSyncCheckRetries      = 30
-	timeBetweenSyncCheckRetries = 1 * time.Second
 
 	beaconSuffixServiceId    = "beacon"
 	validatorSuffixServiceId = "validator"
@@ -85,13 +81,12 @@ func (launcher *LodestarClientLauncher) Launch(
 		return nil, stacktrace.NewError("Expected new Lodestar beacon service to have port with ID '%v', but none was found", httpPortID)
 	}
 
-	restClient := cl_client_rest_client.NewCLClientRESTClient(beaconServiceCtx.GetPrivateIPAddress(), httpPort.GetNumber())
-
-	if err := waitForAvailability(restClient); err != nil {
+	beaconRestClient := cl_client_rest_client.NewCLClientRESTClient(beaconServiceCtx.GetPrivateIPAddress(), httpPort.GetNumber())
+	if err := availability_waiter.WaitForBeaconClientAvailability(beaconRestClient, maxNumHealthcheckRetries, timeBetweenHealthcheckRetries); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred waiting for the new Lodestar beacon node to become available")
 	}
 
-	nodeIdentity, err := restClient.GetNodeIdentity()
+	nodeIdentity, err := beaconRestClient.GetNodeIdentity()
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting the new Lodestar beacon node's identity, which is necessary to retrieve its ENR")
 	}
@@ -110,15 +105,11 @@ func (launcher *LodestarClientLauncher) Launch(
 		return nil, stacktrace.Propagate(err, "An error occurred launching the Lodestar CL validator client with service ID '%v'", serviceId)
 	}
 
-	if err := waitForSync(restClient); err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred waiting for the new Lodestar validator node to become synced with the beacon node")
-	}
-
 	result := cl.NewCLClientContext(
 		nodeIdentity.ENR,
 		beaconServiceCtx.GetPrivateIPAddress(),
 		httpPortNum,
-		restClient,
+		beaconRestClient,
 	)
 
 	return result, nil
@@ -127,7 +118,6 @@ func (launcher *LodestarClientLauncher) Launch(
 // ====================================================================================================
 //                                   Private Helper Methods
 // ====================================================================================================
-
 func getBeaconContainerConfigSupplier(
 		bootnodeContext *cl.CLClientContext, // If this is empty, the node will be launched as a bootnode
 		elClientContext *el.ELClientContext,
@@ -221,9 +211,6 @@ func getValidatorContainerConfigSupplier(
 
 		rootDirpath := path.Join(consensusDataDirpathOnServiceContainer, string(serviceId))
 
-		logrus.Infof("Lodestar Keystore Dirpath: %v", validatorKeysDirpathOnModuleContainer)
-		logrus.Infof("Lodestar Secrets Dirpath: %v", validatorSecretsDirpathOnModuleContainer)
-
 		cmdArgs := []string{
 			"validator",
 			"--rootDir=" + rootDirpath,
@@ -245,36 +232,4 @@ func getValidatorContainerConfigSupplier(
 		return containerConfig, nil
 	}
 	return containerConfigSupplier
-}
-
-
-func waitForAvailability(restClient *cl_client_rest_client.CLClientRESTClient) error {
-	for i := 0; i < maxNumHealthcheckRetries; i++ {
-		_, err := restClient.GetHealth()
-		if err == nil {
-			// TODO check the healthstatus???
-			return nil
-		}
-		time.Sleep(timeBetweenHealthcheckRetries)
-	}
-	return stacktrace.NewError(
-		"Lodestar node didn't become available even after %v retries with %v between retries",
-		maxNumHealthcheckRetries,
-		timeBetweenHealthcheckRetries,
-	)
-}
-
-func waitForSync(restClient *cl_client_rest_client.CLClientRESTClient) error {
-	for i := 0; i < maxNumSyncCheckRetries; i++ {
-		syncingData, err := restClient.GetNodeSyncingData()
-		if err == nil && syncingData.IsSyncing {
-			return nil
-		}
-		time.Sleep(timeBetweenSyncCheckRetries)
-	}
-	return stacktrace.NewError(
-		"Lodestar validator node didn't become syncing with beacon node even after %v retries with %v between retries",
-		maxNumSyncCheckRetries,
-		timeBetweenSyncCheckRetries,
-	)
 }