+++ docs/changelog.md
@@ -1,4 +1,10 @@
 # TBD
+### Features
+* Added Prysm CL (beacon, validator) Launcher
+
+### Changes
+* The `WaitForBeaconClientAvailability` method also checks if the returned status is READY, which means the node is synced
+* Replaced the custom implementation of the availability waiter method in Lodestar Launcher with the `WaitForBeaconClientAvailability` used for other launchers
 
 # 0.2.3
 ### Features